Refactor list quotes * Removed unused imports from all command classes
