+++ src/main/java/seedu/quotesify/commands/delete/DeleteQuoteCommand.java
@@ -1,6 +1,5 @@
 package seedu.quotesify.commands.delete;
 
-import seedu.quotesify.commands.Command;
 import seedu.quotesify.exception.QuotesifyException;
 import seedu.quotesify.lists.ListManager;
 import seedu.quotesify.quote.Quote;
@@ -15,19 +14,19 @@ public class DeleteQuoteCommand extends DeleteCommand {
     }
 
     public void execute(TextUi ui, Storage storage) {
-        QuoteList quotes = (QuoteList) ListManager.getList(ListManager.QUOTE_LIST);
-        deleteQuote(quotes, ui, information);
+        QuoteList quoteList = (QuoteList) ListManager.getList(ListManager.QUOTE_LIST);
+        deleteQuote(quoteList, ui, information);
         storage.save();
     }
 
-    private void deleteQuote(QuoteList quotes, TextUi ui, String information) {
+    private void deleteQuote(QuoteList quoteList, TextUi ui, String information) {
         try {
             if (information.trim().isEmpty()) {
                 throw new QuotesifyException(ERROR_NO_QUOTE_NUMBER);
             }
             int quoteNumber = Integer.parseInt(information.trim()) - 1;
-            Quote quoteToBeDeleted = quotes.getQuote(quoteNumber);
-            quotes.delete(quoteNumber);
+            Quote quoteToBeDeleted = quoteList.getQuote(quoteNumber);
+            quoteList.delete(quoteNumber);
             ui.printDeleteQuote(quoteToBeDeleted.getQuote());
         } catch (NumberFormatException | IndexOutOfBoundsException e) {
             System.out.println(ERROR_INVALID_QUOTE_NUM);
